Fix #282: Allow our add-another feature in admin.
As users reported[0], the add-another plus button is missing when
overriding a (Multiple)ModelChoiceField in a form that is used in Django
admin.

As a (cool) workaround, django-autocomplete-light is now able to use its
own add-another feature in Django admin too, for this kind of cases.

To activate this feature, we just need to set ``add_another_url_name``
Autocomplete's argument, ie.::

    autocomplete_light.register(MyModel,
        add_another_url_name='admin:yourapp_yourmodel_add')

Of course it's a workaround, the idea would be to transparently support
django.contrib.admin.widget.RelatedFieldWidgetWrapper and as you'll see
it's a bit complicated, might not be worth it for now.

[0]
https://github.com/yourlabs/django-autocomplete-light/issues/282#issuecomment-68081452

[1]
http://stackoverflow.com/questions/18602563/django-modelchoicefield-has-no-plus-button