Support Django 1.8 change-link.
This is not my favorite fix, ideally we'd have seen this coming by
building against the master branch and were able to participate to the
elaboration of this admin widget.

So, the widget now has 2 new events it triggers:

- widgetSelectChoice,
- widgetDeselectChoice.

Also, a couple of callbacks were bound to those to support Django 1.8
change-link href management.