Dynamic evaluation of the autocomp box container.
In 2.2.0 series, we want most complete support of Django admin - which means
all edge cases - and it's oldschool overflow: hidden way of clearing floats.

On the other hand, we also want top-notch support for modern CSS patterns such
as the .clearfix which is the most-standard way of clearing floats IMHO and is
of course the way all modern CSS frameworks clear their floats.

By default, this traverses this.input's parents to find the nearest parent
with an 'absolute' or 'fixed' position. This prevents scrolling issues that
we'd have by boldly using <body>.