Do not enforce semicolon
http://jshint.com/docs/options/

There is a lot of FUD about semicolon spread by quite a few people in
the community. The common myths are that semicolons are required all the
time (they are not) and that they are unreliable. JavaScript has rules
about semicolons which are followed by all browsers so it is up to you
to decide whether you should or should not use semicolons in your code.

For more information about semicolons in JavaScript read An Open Letter
to JavaScript Leaders Regarding Semicolons by Isaac Schlueter and
JavaScript Semicolon Insertion.